% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinReg.R
\name{LinReg}
\alias{LinReg}
\title{Simple Linear Regression - Data Annealing SMC.}
\usage{
LinReg(dat, particles = 1000)
}
\arguments{
\item{dat}{A two-column matrix or dataframe containing x and y
values. The data set from Williams (1959) is used as the
default if no data is supplied.}

\item{particles}{An integer specifying the number of particles.}
}
\value{
A \code{list} containing the means and variances of 
the estimated \eqn{alpha}{alpha}, \eqn{beta}{beta} and \eqn{phi}{phi} at each iteration. The
effective sample size at each of the iterations and the logarithm of the estimated model
evidence (normalising constant of the posterior) are also returned.
}
\description{
The \code{LinReg} function provides a simple example for \pkg{RcppSMC} based on
estimating the parameters of a linear regression model using data annealing sequential Monte Carlo.
}
\details{
The \code{LinReg} function implements a simple example based on
estimating the parameters of a linear regression model using data
annealing SMC. Details of the experiment on which the default
data is based can be found in Williams (1959).

The log evidence is -309.56 rounded to two decimal places.
}
\examples{
res <- LinReg(particles=1000)

}
\references{
Williams, E. (1959). Regression analysis. Wiley.
}
\seealso{
\code{\link{LinRegLA}} for an implementation using likelihood annealing SMC.
}
\author{
Adam M. Johansen, Dirk Eddelbuettel and Leah F. South
}
\concept{
static Bayesian annealing
}
\keyword{programming}
