% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinRegLA.R
\name{LinRegLA}
\alias{LinRegLA}
\title{Simple Linear Regression - Likelihood Annealing SMC}
\usage{
LinRegLA(dat, temperatures, particles = 1000)
}
\arguments{
\item{dat}{A two-column matrix or dataframe containing x and y
values. The data set from Williams (1959) is used as the
default if no data is supplied.}

\item{temperatures}{Here the targets are defined by \eqn{P(y|\theta)^{\gamma_t}P(\theta)}
where \eqn{0=\gamma_0\le \ldots \le \gamma_T = 1} can be referred to
as the temperatures, \eqn{P(y|\theta)} is the likelihood and \eqn{P(\theta)}
is the prior. An example temperature schedule is used if one is not specified.}

\item{particles}{An integer specifying the number of particles.}
}
\value{
A \code{list} containing the means and variances of 
the estimated \eqn{alpha}{alpha}, \eqn{beta}{beta} and \eqn{phi}{phi} at each iteration. The
effective sample size at each of the iterations and the logarithm of the estimated model
evidence (normalising constant of the posterior) are also returned.
}
\description{
The \code{LinReg} function provides a simple example for \pkg{RcppSMC} based on
estimating the parameters of a linear regression model using likelihood annealing sequential Monte Carlo.
}
\details{
The \code{LinRegLA} function implements a simple example based on
estimating the parameters of a linear regression model using likelihood
annealing SMC. The details relevant to the default choices can be found
on page 49 of \url{https://eprints.qut.edu.au/101729/7/101729.pdf}.
Details of the original experiment can be found in Williams (1959).

The log evidence is -309.56 rounded to two decimal places.
}
\examples{
res <- LinRegLA(particles=1000)

}
\references{
Williams, E. (1959). Regression analysis. Wiley.
}
\seealso{
\code{\link{LinReg}} for an implementation using data annealing SMC.
}
\author{
Adam M. Johansen, Dirk Eddelbuettel and Leah F. South
}
\concept{
static Bayesian annealing path
}
\keyword{programming}
