% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{pfdata}
\alias{pfdata}
\title{Particle filtering data}
\format{Add my format.}
\source{
Add my source.
}
\usage{
pfdata
}
\description{
Add my description. The variables are as follows:
}
\details{
\itemize{
  \item item1
}
}
\keyword{datasets}
