% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{radiata}
\alias{radiata}
\title{Radiata pine regression dataset}
\format{A data frame with this many rows and this many variables}
\source{
Williams, E. (1959). Regression analysis. Wiley.
}
\usage{
radiata
}
\description{
Add my description. The variables are as follows:
}
\details{
\itemize{
  \item var1 (range a--b)
}
}
\keyword{datasets}
