% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppSMC.R
\docType{package}
\name{RcppSMC-package}
\alias{RcppSMC}
\alias{RcppSMC-package}
\title{A package for general implementation of SMC algorithms}
\description{
RcppSMC can be used for algorithms ranging from simple particle filters
to complicated SMC algorithms
}
\references{
Add link to SMCTC paper and other references
}
