% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{pfdata}
\alias{pfdata}
\title{Particle filtering data}
\format{A data frame with 100 rows and two variables:
\describe{
 \item{x}{Observed x positions}
 \item{y}{Observed y positions}
}}
\usage{
pfdata
}
\description{
This is the default data set for the example in \code{pfLineartBS}.
The dataset was simulated from the linear model with t-distibuted
innovations described in Section 5.1 of Johansen (2009).
}
\references{
A. M. Johansen. SMCTC: Sequential Monte Carlo in C++.
Journal of Statistical Software, 30(6):1-41, April
2009. \url{http://www.jstatsoft.org/v30/i06/paper}
}
\seealso{
\code{\link{pfLineartBS}}
}
\keyword{datasets}
