\docType{data}
\name{radiata}
\alias{radiata}
\title{Radiata pine dataset (linear regression example)}
\format{A data frame with 42 rows and three variables:
\describe{
 \item{y}{Maximum compression strength (response) in pounds per square inch}
 \item{x1}{Density (predictor 1) in pounds per cubic foot}
 \item{x2}{Adjusted density (predictor 2) in pounds per cubic foot}
}}
\source{
E. Williams. Regression analysis. Wiley, 1959.
}
\usage{
radiata
}
\description{
This dataset was originally presented in Table 5.1 of Williams (1959) where
two non-nested linear regression models were considered.
}
\keyword{datasets}
